move all dash files to dash component to support things other than socket
