+++ src/components/external-dash/UserInfo.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { User } from "../types/User";
+import { User } from "./types/User";
 
 interface Props {
   userData: User;
