+++ src/components/external-dash/types/ClientQuery.ts
